backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
prisma==0.15.0
google-auth==2.35.0
httpx==0.27.0
pydantic==2.9.0
requests==2.31.0
email-validator==2.2.0

